backend/requirements.txt

fastapi
pydantic
typing-extensions
uvicorn
psycopg2-binary
python-dotenv
dotenv
supabase
boto3